Rename example to custom flag
