+++ apps/cookbook/src/app/examples/header-example/header-example.module.ts
@@ -14,7 +14,7 @@ import { HeaderExampleValueComponent } from './examples/value';
 import { HeaderExampleCombinedComponent } from './examples/combined';
 import { HeaderExampleCustomSectionComponent } from './examples/custom-section';
 import { HeaderExampleTitleScalingComponent } from './examples/title-scaling';
-import { HeaderExampleCustomFlagComponent } from './examples/custom-content-flag';
+import { HeaderExampleCustomFlagComponent } from './examples/custom-flag';
 
 const COMPONENT_DECLARATIONS = [
   HeaderExampleComponent,
